pyproject.toml
[tool.poetry]
name = "notion-to-slack"
version = "0.1.0"
description = ""
authors = ["maskai"]
readme = "README.md"
packages = [{include = "notion_to_slack"}]

[tool.poetry.dependencies]
python = "^3.9"
python-dotenv = "^1.0.0"
schedule = "^1.2.0"

[tool.poetry.group.dev.dependencies]
requests = "^2.31.0"
slack-sdk = "^3.22.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
